fix clippy lints

+++ src/threadstat_message.rs
@@ -1,5 +1,5 @@
 mod raw {
-    #![allow(non_upper_case_globals, non_camel_case_types, non_snake_case, dead_code)]
+    #![allow(non_upper_case_globals, non_camel_case_types, non_snake_case, dead_code, clippy::pedantic)]
     include!(concat!(env!("OUT_DIR"), "/shared_def_bindings.rs"));
 }
 
@@ -26,7 +26,7 @@ pub struct ThreadstatMessage {
 }
 
 #[derive(Debug)]
-pub struct UnknownFlag(pub i32);
+pub struct UnknownFlag(pub u32);
 
 impl std::fmt::Display for UnknownFlag {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
@@ -41,8 +41,8 @@ impl TryFrom<message> for ThreadstatMessage {
 
     fn try_from(msg: message) -> Result<Self, Self::Error> {
         let flag = match msg.flags {
-            x if x == raw::MESSAGE_FLAG_ADD_PROCESS as i32 => MessageFlag::AddProcess,
-            x if x == raw::MESSAGE_FLAG_REMOVE_PROCESS as i32 => MessageFlag::RemoveProcess,
+            x if x == raw::MESSAGE_FLAG_ADD_PROCESS => MessageFlag::AddProcess,
+            x if x == raw::MESSAGE_FLAG_REMOVE_PROCESS => MessageFlag::RemoveProcess,
             x => return Err(UnknownFlag(x)),
         };
         Ok(ThreadstatMessage { tid: msg.tid, flag })
